pyproject.toml
[project]
name = "udemy-my_work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "wat-inspector>=0.4.3",
]
